requirements.txt
streamlit==1.40.0
yfinance==0.2.50
pandas==2.2.3
numpy==1.26.4
scikit-learn==1.5.2
xgboost==2.1.3
tensorflow==2.18.0
ta==0.11.0
plotly==5.24.1
openpyxl==3.1.5
sqlalchemy==2.0.36
apscheduler==3.11.0
python-dotenv==1.0.1
loguru==0.7.3
tenacity==9.0.0
joblib==1.4.2
matplotlib==3.9.2
seaborn==0.13.2
kaleido==0.2.1
email-validator==2.2.0
pytest==8.3.4
pytest-cov==6.0.0
